requirements.txt
asgiref==3.7.2
boto3==1.34.65
botocore==1.34.65
coverage==7.4.1
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
Markdown==3.5.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2023.4
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.2.1
